backend/requirements/base.txt
# Django Core
Django==5.0.0
djangorestframework==3.14.0
django-cors-headers==4.3.0
django-filter==23.5

# Authentication
djangorestframework-simplejwt==5.3.0

# Database
psycopg2-binary==2.9.9

# Async & WebSocket
channels==4.0.0
channels-redis==4.2.0
daphne==4.0.0

# Celery
celery==5.3.0
redis==5.0.0

# File Storage
django-storages==1.14.2
boto3==1.34.0
pillow==10.2.0

# Search
elasticsearch==8.11.0
django-elasticsearch-dsl==8.0

# API Documentation
drf-yasg==1.21.7

# Utilities
python-dotenv==1.0.0
pytz==2023.3
